requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.68.1
anyio==4.10.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
bitsandbytes==0.47.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.3
click==8.2.1
cohere==5.18.0
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cut-cross-entropy==25.1.1
dataclasses-json==0.6.7
datasets==3.6.0
dbus-python==1.2.18
decorator==5.2.1
diffusers==0.35.1
dill==0.3.8
diskcache==5.6.3
distro==1.7.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
executing==2.2.1
faiss-gpu==1.7.2
fastavro==1.12.0
filelock==3.19.1
-e git+https://github.com/glebkkkle/Find_yourself.git@331fde1a7804ead640731708b20e7b0e8f694657#egg=find_yourself
firebase_admin==7.1.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hf_transfer==0.1.9
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.5
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
langchain==0.3.27
langchain-anthropic==0.3.20
langchain-cohere==0.4.6
langchain-community==0.3.29
langchain-core==0.3.76
langchain-huggingface==0.3.1
langchain-ollama==0.3.8
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.27
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama_cpp_python==0.3.16
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.6.0
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
ollama==0.6.0
openai==1.107.2
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.10.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2022.1
PyYAML==5.4.1
redis==6.4.0
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
sentence-transformers==5.1.0
sentencepiece==0.2.1
service-identity==18.1.0
shtab==1.7.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.50.0
sympy==1.14.0
systemd-python==234
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0+cu126
torchao==0.13.0
torchaudio==2.8.0+cu126
torchvision==0.23.0+cu126
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
triton==3.4.0
trl==0.23.0
Twisted==22.1.0
typeguard==4.4.4
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
ufw==0.36.1
unattended-upgrades==0.1
unsloth==2025.9.5
unsloth_zoo==2025.9.6
uritemplate==4.2.0
urllib3==2.5.0
utils==1.0.2
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
xformers==0.0.32.post2
xxhash==3.5.0
yarl==1.20.1
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.25.0
